requirements.txt
openai>=1.35.0
gspread>=6.0.0
google-auth>=2.29.0
google-api-python-client>=2.137.0
python-dotenv>=1.0.1
typer>=0.12.3
rich>=13.7.1
tenacity>=8.2.3
pydantic>=2.7.0

